exception: MissingInputException

+++ src/main/java/facades/PersonFacade.java
@@ -4,6 +4,7 @@ import dtos.PersonDTO;
 
 import dtos.PersonsDTO;
 import entities.Person;
+import errorhandling.MissingInputException;
 import errorhandling.PersonNotFoundException;
 
 import javax.persistence.EntityManager;
@@ -34,8 +35,13 @@ public class PersonFacade implements IPersonFacade {
     }
 
     @Override
-    public PersonDTO addPerson(String fName, String lName, String phone) {
+    public PersonDTO addPerson(String fName, String lName, String phone) throws MissingInputException {
         Person person = new Person(fName, lName, phone);
+
+        if ((fName == null) || (lName == null)) {
+            throw new MissingInputException(400, "First Name and/or Last Name is missing");
+        }
+
         EntityManager em = getEntityManager();
         try {
             em.getTransaction().begin();
@@ -48,9 +54,12 @@ public class PersonFacade implements IPersonFacade {
     }
 
     @Override
-    public PersonDTO deletePerson(int id) {
+    public PersonDTO deletePerson(int id) throws PersonNotFoundException {
         EntityManager em = getEntityManager();
         Person person = em.find(Person.class, id);
+        if (person == null) {
+            throw new PersonNotFoundException(404, "Could not delete, provided id does not exist");
+        }
         try {
             em.getTransaction().begin();
             em.remove(person);
@@ -66,7 +75,7 @@ public class PersonFacade implements IPersonFacade {
         EntityManager em = getEntityManager();
         Person person = em.find(Person.class, id);
         if (person == null) {
-            throw new PersonNotFoundException(404, "message: No person with provided id found");
+            throw new PersonNotFoundException(404, "No person with provided id found");
         }
         return new PersonDTO(person);
     }
@@ -80,10 +89,18 @@ public class PersonFacade implements IPersonFacade {
     }
 
     @Override
-    public PersonDTO editPerson(PersonDTO pDTO) {
+    public PersonDTO editPerson(PersonDTO pDTO) throws PersonNotFoundException, MissingInputException {
         EntityManager em = getEntityManager();
 
         Person person = em.find(Person.class, pDTO.getId());
+        if (person == null) {
+            throw new PersonNotFoundException(404, "Could not update, provided id does not exist");
+        }
+
+        if ((pDTO.getFirstName() == null) || (pDTO.getLastName() == null)) {
+            throw new MissingInputException(400, "First Name and/or Last Name is missing");
+        }
+
         person.setFirstName(pDTO.getFirstName());
         person.setLastName(pDTO.getLastName());
         person.setPhone(pDTO.getPhone());
@@ -91,6 +108,7 @@ public class PersonFacade implements IPersonFacade {
 
         try {
             em.getTransaction().begin();
+            //em.merge(person); <-- Koden virker uden?
             em.getTransaction().commit();
         } finally {
             em.close();
